refactor(web): inline Ollama engine into unified-engine
Ollama HTTP and URL resolution already live in me-ai-core; remove
ollama-engine.ts and use createOllamaEngine() beside the cloud adapter.

Made-with: Cursor
